refactor(core)!: move phrases api to well-known path
