+++ .changeset-staged/stupid-jokes-brush.md
@@ -1,0 +1,6 @@
+---
+"@logto/core": major
+"@logto/ui": major
+---
+
+**💥 BREAKING CHANGE 💥** Move `/api/phrase` API to `/api/.well-known/phrases`
